Update 99-further-reading-1.md
Hi, thanks for making this presentation available on the web. It was actually of great help to me in trying to share with others in my business how event sourcing works and the ideas behind it - I hope that's ok. Thank you :) :)

I especially appreciate it because there seems very few places available that the principles required are spelled out. And so, after several years building event sourced systems, I've spent the last two collating my learnings and articles, and also interviewed several other companies that had taken this approach. I put the results of my work into a youtube talk, as I felt it helped to have the narrative available, and which I hope helps bring together the principles of each component, so that we get more common & shared understanding and can move forward faster. And I hoped you'd consider including it as a link at the end of your talk, as maybe a next step for anyone wanting more detail after listening to yours, or to build their own system. And I hope you'll consider too turning your talk into a youtube one too 👍 

Once again, many thanks for compiling this talk and making it available to everyone!